move unit/conftest.py one level lower
