+++ tests/unit/test_dry_run.py
@@ -6,8 +6,7 @@ import tempfile
 from pathlib import Path
 
 from tests.utils import defaultdict_to_standard_types
-from .conftest import RINGTEST_DIR, NGV_DIR
-from ..conftest import PLATFORM_SYSTEM
+from ..conftest import PLATFORM_SYSTEM, RINGTEST_DIR, NGV_DIR
 
 TMP_FOLDER = tempfile.mkdtemp()
