+++ tox.ini
@@ -45,7 +45,7 @@ setenv =
     NEURON_INIT_MPI=1
 commands =
     python -c "import os; print(os.environ.get('HOC_LIBRARY_PATH', ''))"
-    pytest {posargs:tests/mpi}
+    pytest -x {posargs:tests/mpi}
 
 [testenv:integration-e2e]
 # Please you need to spack load neurodamus-models model=neocortex +coreneuron or use an equivalent spack virtual env
